Refactor: 일별 판매량 이벤트 리스너 수정
