--- src/main/java/kr/hhplus/be/server/application/order/OrderCompletedEvent.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.application.order;
-
-import kr.hhplus.be.server.domain.order.OrderItem;
-
-import java.util.List;
-
-public record OrderCompletedEvent(List<OrderItem> items) {}
