+++ src/main/java/kr/hhplus/be/server/domain/product/stats/event/ProductSalesDailyEventListener.java
@@ -1,5 +1,6 @@
-package kr.hhplus.be.server.application.order;
+package kr.hhplus.be.server.domain.product.stats.event;
 
+import kr.hhplus.be.server.application.event.ProductSalesDailyEvent;
 import kr.hhplus.be.server.domain.product.stats.ProductSalesService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
@@ -9,13 +10,13 @@ import org.springframework.transaction.event.TransactionalEventListener;
 
 @RequiredArgsConstructor
 @Component
-public class OrderCompletedEventListener {
+public class ProductSalesDailyEventListener {
 
     private final ProductSalesService productSalesService;
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     @Async
-    public void handle(OrderCompletedEvent event) {
+    public void handle(ProductSalesDailyEvent event) {
         productSalesService.increaseDailySales(event.items());
     }
 }
